refactor: refactor dog-profile-trials-laaja-page.tsx to smaller components ref bej-125
